+++ src/HI.Sample/ConsumerCreateVerifiedIHIClientSample.cs
@@ -13,12 +13,12 @@
  */
 
 using System;
+using System.Net;
 using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.Security.Cryptography.X509Certificates;
 using nehta.mcaR40.CreateVerifiedIHI;
 using Nehta.VendorLibrary.Common;
-
 
 namespace Nehta.VendorLibrary.HI.Sample
 {
@@ -38,6 +38,8 @@ namespace Nehta.VendorLibrary.HI.Sample
             // ------------------------------------------------------------------------------
             // Set up
             // ------------------------------------------------------------------------------
+
+            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
 
             // Obtain the certificate by serial number
             X509Certificate2 tlsCert = X509CertificateUtil.GetCertificate(
@@ -77,7 +79,7 @@ namespace Nehta.VendorLibrary.HI.Sample
             QualifiedId hpio = new QualifiedId()
             {
                 id = "HPIO",                                              // HPIO internal to your system
-                qualifier = "http://<anything>/id/<anything>/hpio/1.0"    // Eg: http://ns.yourcompany.com.au/id/yoursoftware/userid/1.0
+                qualifier = "http://ns.electronichealth.net.au/id/hi/hpio/1.0"
             };
 
             // ------------------------------------------------------------------------------