+++ src/HI.Sample/ConsumerSearchIHIBatchSyncClientSample.cs
@@ -16,6 +16,7 @@ using System;
 using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.Collections.Generic;
+using System.Net;
 using System.Security.Cryptography.X509Certificates;
 using Nehta.VendorLibrary.HI.Common;
 using nehta.mcaR3.ConsumerSearchIHIBatchSync;
@@ -39,6 +40,8 @@ namespace Nehta.VendorLibrary.HI.Sample
             // ------------------------------------------------------------------------------
             // Set up
             // ------------------------------------------------------------------------------
+
+            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
 
             // Obtain the certificate by serial number
             X509Certificate2 tlsCert = X509CertificateUtil.GetCertificate(
@@ -78,7 +81,7 @@ namespace Nehta.VendorLibrary.HI.Sample
             QualifiedId hpio = new QualifiedId()
             {
                 id = "HPIO",                                              // HPIO internal to your system
-                qualifier = "http://<anything>/id/<anything>/hpio/1.0"    // Eg: http://ns.yourcompany.com.au/id/yoursoftware/userid/1.0
+                qualifier = "http://ns.electronichealth.net.au/id/hi/hpio/1.0"
             };
 
             // ------------------------------------------------------------------------------